refactor(babel): :art:Add babel build script to improve performace
