+++ scripts/build.js
@@ -1,0 +1,8 @@
+/**
+ * @since 2017-03-18 10:55:28
+ * @author vivaxy
+ */
+
+require('babel-polyfill');
+require('babel-register')();
+require('./build/index');
