+++ src/commands/config.js
