+++ src/config/config.js
@@ -1,4 +1,9 @@
-{
+/**
+ * @since 2016-11-16 10:17
+ * @author vivaxy
+ */
+
+export default {
     "scaffold": {
         "vivaxy/gt-react-scaffold": {
             "repo": "https://github.com/vivaxy/gt-react-scaffold.git",
