+++ src/file/directoryExists.js
