+++ src/git/checkGitRepository.js
