+++ src/lib/configManager.js
