+++ src/lib/getInfoFromShell.js
