+++ src/lib/updateScaffoldStat.js
