+++ src/presets/copyFiles.js
