+++ src/presets/removeFiles.js
